pyproject.toml
[tool.poetry]
name = "unfold-studio"
version = "0.5.0"
description = "Web application for interactive storytelling"
authors = ["Chris Proctor <chris@chrisproctor.net>"]
license = "MIT"
readme = "README.md"
packages = [{include = "unfold_studio"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
django-storages = "^1.14.3"
django-reversion = "^5.0.12"
django-cors-headers = "^4.4.0"
django-extensions = "^3.2.3"
social-auth-app-django = "^5.4.1"
django-qr-code = "^4.1.0"
arrow = "^1.3.0"
psycopg = "^3.1.19"


[tool.poetry.group.docs.dependencies]
sphinx = "^7.3.7"
sphinx-rtd-theme = "^2.0.0"
sphinxcontrib-bibtex = "^2.6.2"
sphinxcontrib-websupport = "^1.2.7"


[tool.poetry.group.admin.dependencies]
tabulate = "^0.9.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
